Refactor: Import schematic and blog types from schemas
This commit updates the import path for the `Schematic` and `Blog` types. It now imports them from the schema files instead of the old location. This change ensures that the components use the correct type definitions.
