+++ src/components/features/schematics/UserSchematicList.tsx
@@ -17,7 +17,7 @@ import { useLoggedUser } from '@/api/context/loggedUser/loggedUserContext.tsx';
 import { useThemeStore } from '@/api/stores/themeStore.tsx';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar.tsx';
 import MinecraftIcon from '@/components/utility/MinecraftIcon.tsx';
-import {Schematic} from "@/types";
+import {Schematic} from "@/schemas/schematic.schema.tsx";
 
 const UserSchematicList = () => {
   const { isDarkMode } = useThemeStore();
